+++ Pinnacle/src/main/java/com/cfive/pinnacle/entity/NoticeReceive.java
@@ -57,7 +57,7 @@ public class NoticeReceive implements Serializable {
 
     @TableField("deleted")
     @TableLogic
-    private Integer deleted;
+    private Long deleted;
 
     @TableField("version")
     @Version
