+++ Pinnacle/src/main/java/com/cfive/pinnacle/entity/NoticeType.java
@@ -49,7 +49,7 @@ public class NoticeType implements Serializable {
 
     @TableField("deleted")
     @TableLogic
-    private Integer deleted;
+    private Long deleted;
 
     @TableField("version")
     @Version
